fix some tests

+++ tests/cn/division_casting.c
@@ -5,8 +5,8 @@ Important: (1) signed integers must be non-negative to convert to unsigned (2) i
 is unsigned, the result will be unsigned, so any signed values must be non-negative. */
 
 unsigned int division (unsigned int x, int y)
-/*@ requires y > 0i32;
-    ensures return == x/(u32)y; @*/
+/*@ requires y > 0;
+    ensures return == x/y; @*/
 {
     return x/y;
 }
